refine and fix stylecop
