+++ solutions/Virtual-Assistant/src/csharp/skills/calendarskill/Dialogs/NextMeeting/NextMeetingDialog.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Collections.Specialized;
 using System.Threading;
 using System.Threading.Tasks;
@@ -11,7 +12,6 @@ using Microsoft.Bot.Builder;
 using Microsoft.Bot.Builder.Dialogs;
 using Microsoft.Bot.Solutions.Extensions;
 using Microsoft.Bot.Solutions.Skills;
-using System;
 using Microsoft.Bot.Solutions.Util;
 
 namespace CalendarSkill
