+++ solutions/Virtual-Assistant/src/csharp/skills/tests/CalendarSkillTest/API/Fakes/MockBaseClient/MockBaseServiceClient.cs
@@ -5,7 +5,7 @@ using CalendarSkill.Models;
 using CalendarSkill.ServiceClients;
 using Moq;
 
-namespace CalendarSkillTest.API.Fakes
+namespace CalendarSkillTest.API.Fakes.MockBaseClient
 {
     public static class MockBaseServiceClient
     {
