+++ solutions/Virtual-Assistant/src/csharp/skills/tests/CalendarSkillTest/API/UserServiceMSGraphTests.cs
@@ -1,0 +1,58 @@
+﻿using System.Collections.Generic;
+using System.Threading.Tasks;
+using CalendarSkill.Models;
+using CalendarSkill.ServiceClients;
+using CalendarSkill.ServiceClients.MSGraphAPI;
+using CalendarSkillTest.API.Fakes.MockMSGraphClient;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace CalendarSkillTest.API
+{
+    [TestClass]
+    public class UserServiceMSGraphTests
+    {
+        private static IUserService userService;
+
+        [ClassInitialize]
+        public static void ClassInit(TestContext context)
+        {
+            userService = new UserService(new MSGraphUserService(MockMSGraphUserClient.GetUserService()));
+        }
+
+        [ClassCleanup]
+        public static void ClassCleanup()
+        {
+        }
+
+        [TestInitialize]
+        public void TestInit()
+        {
+        }
+
+        [TestCleanup]
+        public void TestCleanup()
+        {
+        }
+
+        [TestMethod]
+        public async Task GetPeopleAsyncTest()
+        {
+            List<PersonModel> result = await userService.GetPeopleAsync("Doe");
+            Assert.IsTrue(result.Count == 2);
+        }
+
+        [TestMethod]
+        public async Task GetUserAsyncTest()
+        {
+            List<PersonModel> result = await userService.GetUserAsync("Doe");
+            Assert.IsTrue(result.Count == 2);
+        }
+
+        [TestMethod]
+        public async Task GetContactsAsyncTest()
+        {
+            List<PersonModel> result = await userService.GetContactsAsync("Doe");
+            Assert.IsTrue(result.Count == 2);
+        }
+    }
+}
