--- solutions/Virtual-Assistant/src/csharp/skills/tests/calendarskilltest/API/UserServiceTests.cs
@@ -1,160 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using CalendarSkill;
-using CalendarSkill.Models;
-using CalendarSkill.ServiceClients;
-using CalendarSkill.ServiceClients.GoogleAPI;
-using CalendarSkill.ServiceClients.MSGraphAPI;
-using CalendarSkillTest.API.Fakes;
-using Microsoft.Bot.Solutions.Skills;
-using Microsoft.Graph;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-namespace CalendarSkillTest.API
-{
-    [TestClass]
-    public class UserBaseServiceTests
-    {
-        public static IUserService userService;
-
-        [ClassInitialize]
-        public static void ClassInit(TestContext context)
-        {
-            userService = new UserService(MockBaseUserClient.GetUserService());
-        }
-
-        [ClassCleanup]
-        public static void ClassCleanup()
-        {
-        }
-
-        [TestInitialize]
-        public void TestInit()
-        {
-        }
-
-        [TestCleanup]
-        public void TestCleanup()
-        {
-        }
-
-        [TestMethod]
-        public async Task GetPeopleAsyncTest()
-        {
-            List<PersonModel> result = await userService.GetPeopleAsync("Doe");
-            Assert.IsTrue(result.Count == 0);
-        }
-
-        [TestMethod]
-        public async Task GetUserAsyncTest()
-        {
-            List<PersonModel> result = await userService.GetUserAsync("Doe");
-            Assert.IsTrue(result.Count == 0);
-        }
-
-        [TestMethod]
-        public async Task GetContactsAsyncTest()
-        {
-            List<PersonModel> result = await userService.GetContactsAsync("Doe");
-            Assert.IsTrue(result.Count == 0);
-
-        }
-    }
-
-    [TestClass]
-    public class UserMSGraphServiceTests
-    {
-        public static IUserService userService;
-
-        [ClassInitialize]
-        public static void ClassInit(TestContext context)
-        {
-            userService = new UserService(new MSGraphUserService(MockMSGraphUserClient.GetUserService()));
-        }
-
-        [ClassCleanup]
-        public static void ClassCleanup()
-        {
-        }
-
-        [TestInitialize]
-        public void TestInit()
-        {
-        }
-
-        [TestCleanup]
-        public void TestCleanup()
-        {
-        }
-
-        [TestMethod]
-        public async Task GetPeopleAsyncTest()
-        {
-            List<PersonModel> result = await userService.GetPeopleAsync("Doe");
-            Assert.IsTrue(result.Count == 2);
-        }
-
-        [TestMethod]
-        public async Task GetUserAsyncTest()
-        {
-            List<PersonModel> result = await userService.GetUserAsync("Doe");
-            Assert.IsTrue(result.Count == 2);
-        }
-
-        [TestMethod]
-        public async Task GetContactsAsyncTest()
-        {
-            List<PersonModel> result = await userService.GetContactsAsync("Doe");
-            Assert.IsTrue(result.Count == 2);
-        }
-    }
-
-    [TestClass]
-    public class UserGoogleServiceTests
-    {
-        public static IUserService userService;
-
-        [ClassInitialize]
-        public static void ClassInit(TestContext context)
-        {
-            userService = new UserService(new GooglePeopleService(MockGoogleUserClient.GetPeopleService()));
-        }
-
-        [ClassCleanup]
-        public static void ClassCleanup()
-        {
-        }
-
-        [TestInitialize]
-        public void TestInit()
-        {
-        }
-
-        [TestCleanup]
-        public void TestCleanup()
-        {
-        }
-
-        [TestMethod]
-        public async Task GetPeopleAsyncTest()
-        {
-            List<PersonModel> result = await userService.GetPeopleAsync("Doe");
-            Assert.IsTrue(result.Count == 2);
-        }
-
-        [TestMethod]
-        public async Task GetUserAsyncTest()
-        {
-            List<PersonModel> result = await userService.GetUserAsync("Doe");
-            Assert.IsTrue(result.Count == 0);
-        }
-
-        [TestMethod]
-        public async Task GetContactsAsyncTest()
-        {
-            List<PersonModel> result = await userService.GetContactsAsync("Doe");
-            Assert.IsTrue(result.Count == 0);
-        }
-    }
-}
