+++ solutions/Virtual-Assistant/src/csharp/skills/tests/calendarskilltest/Flow/CreateCalendarFlowTests.cs
@@ -360,7 +360,7 @@ namespace CalendarSkillTest.Flow
 
                 var meetingCardJsonString = ((Newtonsoft.Json.Linq.JObject)messageActivity.Attachments[0].Content).ToString();
                 var meetingCard = JsonConvert.DeserializeObject<MeetingAdaptiveCard>(meetingCardJsonString);
-                var meetingInfoList = meetingCard.body[0].items[1].text.Split("\n");
+                var meetingInfoList = meetingCard.Bodies[0].Items[1].Text.Split("\n");
                 var dateString = meetingInfoList[1];
                 CultureInfo cultureInfo = (CultureInfo)CultureInfo.CurrentUICulture.Clone();
                 cultureInfo.DateTimeFormat.DateSeparator = "-";
